updates to negamax
updates. solves instantly for black (negamaxboolean(1)) but not for
white (negamaxboolean(2)) even after waiting 20 seconds. doesn't even
hit the end of the game (to print the colour of the winner). missing
something obvious. following the first easy example with 9 moves left on
board, should be able to solve within 1 second for white.